c5_seguridad_autenticacion/app/schemas/auth_schema.py
from typing import Union
from pydantic import BaseModel


class TokenSchema(BaseModel):
    access_token: str
    token_type: str


class TokenUserSchema(BaseModel):
    email: Union[str, None] = None
