c5_seguridad_autenticacion/requirements.txt
fastapi[all]
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
sqlalchemy
